// src/features/tasks-v2/types.ts
import { paths } from '@/generated/api-schema'
import { UserProfile } from '@/features/me/types'

export interface Assignment {
  assignmentId: number
  taskId: number
  recipientType: string
  assignedAt: string
  dueAt: string | null
  completedAt: string | null
  status: string
  note: string
  assignedByUser: UserProfile
  completedByUser: UserProfile | null
  recipientUser: UserProfile
  recipientId: number
}

export interface Attachment {
  id: number
  filePath: string
  fileName: string
  fileSize: number
  createdAt: string
  uploadedBy: UserProfile
  sharedCount: number | null
}

export interface Task {
  id: number
  title: string
  content: string
  instructions: string
  notes: string | null
  createdAt: string
  updatedAt: string
  createdByUser: UserProfile
  assignments: Assignment[]
  status: 'OPEN' | 'COMPLETE' | 'IN_PROGRESS' | 'TO_DO'
  priority: 'URGENT' | 'NORMAL' | 'HIGH' | 'MEDIUM' | 'LOW'
  parentId: number | null
  subtasks: Task[]
  hierarchyLevel: number | null
  attachments: Attachment[]
}

export interface StatusGroupProps {
  status: 'COMPLETE' | 'IN_PROGRESS' | 'TO_DO' | 'OPEN'
  tasks: Task[]
  expanded: boolean
  onToggle: () => void
  onTaskClick?: (taskId: number) => void
}

export interface TaskRowProps {
  task: Task
  level?: number
  expanded?: boolean
  onToggle?: () => void
}

export type TaskFilters = Omit<Task, 'assignments' | 'documents'>

export type TaskAssignment = Required<Task>['assignments'][number]

export type TaskStatus = NonNullable<Task['status']>

export type TaskAssignmentStatus = NonNullable<TaskAssignment['status']>

export type TaskAssignmentComment = Required<
  NonNullable<
    paths['/api/assignments/{id}/comments']['get']['responses']['200']['content']['*/*']['data']
  >
>[number]

export type TaskFilterTypes = 'created' | 'assigned' | 'received'